refactor(stages): remove the concept of sub-stages
