+++ .storybook/tailwind.config.cjs
@@ -7,6 +7,7 @@ module.exports = {
   content: [
     './packages/login-widget/src/**/*.{html,js,svelte,ts}',
     './core/**/*.{html,js,svelte,ts}',
+    './apps/**/*.{html,js,svelte,ts}',
   ],
   darkMode: 'class',
   presets: [require('../themes/default/config.cjs')],
